change of parameters
Parameters change. Should find circles better. There may appear some
misteakes if a range is big. If a circle is slightly out of a range it
can be approximated by smaller one. I observed it only for biggest
circles.  When we look for a hole of a specific size it shouldn't be a
problem.