Angle Veryfication Class
It detect 10 best corners. Our element has 10 specific corners. Then
they are divided to the left side and to the right  side. Then it looks
for 2 corners we want to compare from the left side. Class works on
blender file but if will detect diffrent 10 corners it may not work.