Material veryfication added
Function detect biggest bright field and compare sides of bounding
rectangle. Proportion between 1 and 2 is characteristik for plastik and
over two for metal. In case when big part of object is reflected and
bounding rectangle is almoust whole object, it will be also detect as
metal. It detects fields in the environment so It needs to be upgraded.